dlib scaling.

+++ dlib-test/fhog-test.cpp
@@ -34,7 +34,7 @@ int main()
             cout << text.length() << endl;
             const char* data = text.c_str();
 
-            const int* rects = fhog_svm_det("IMG_20170124_205102_1.jpg", "object_detector.svm", text.length());
+            const double* rects = fhog_svm_det("IMG_20170124_205102_1.jpg", "object_detector.svm", text.length());
             cout << rects[0] << endl;
             cout << rects[1] << endl;
             cout << rects[2] << endl;
